chapter_002Collections. Pro/src/main/java/ru/job4j/generic/store/UserStore.java
package ru.job4j.generic.store;

/**
 *Class UserStore реализация хранилища для пользователя.
 *@author Chupin Pavel
 *@since 30.06.2021
 */

public class UserStore implements Store<User> {

    private final Store<User> store = new MemStore<>();

    @Override
    public void add(User model) {
        store.add(model);
    }

    @Override
    public boolean replace(String id, User model) {
        return store.replace(id, model);
    }

    @Override
    public boolean delete(String id) {
        return store.delete(id);
    }

    @Override
    public User findById(String id) {
        return store.findById(id);
    }
}
